% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morandi_palette.R
\name{morandi}
\alias{morandi}
\title{Morandi Color Palette for R Plot
Morandi colors refer to a muted and pale color palette,
which is not bright as if covered with a layer of gray tone.
Morandi colors has rich connotation without a tendency to show off,
releasing the soothing elegance.}
\usage{
morandi(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Palette name for plotting two variables:
\code{Two_Variables1}, \code{Two_Variables2},
three variables: \code{Three_Variables1}, \code{Three_Variables2}
multivariables: \code{WildDream}, \code{SweetDream}, \code{MistyForest},
\code{Melancholy}, \code{Aurora}, \code{Cava}.}

\item{n}{number of colors.}

\item{type}{"Continuous" or "discrete".}
}
\value{
A vector of colors
}
\description{
Morandi Color Palette for R Plot
Morandi colors refer to a muted and pale color palette,
which is not bright as if covered with a layer of gray tone.
Morandi colors has rich connotation without a tendency to show off,
releasing the soothing elegance.
}
\examples{
morandi("Two_Variables2")
morandi("Three_Variables2")
morandi("Cava")
morandi("Unicorn")
# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
palette <- morandi(100, name = "Melancholy", type = "continuous")
image(volcano, col = palette)
}
\keyword{colors}
