% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morandi_colors.R
\name{morandi_diy}
\alias{morandi_diy}
\title{DIY your own morandi palette}
\usage{
morandi_diy(my_colors, n, type = c("discrete", "continuous"))
}
\arguments{
\item{my_colors}{a numeric vector}

\item{n}{number of colors}

\item{type}{"discrete" or "continuous"}
}
\value{
a color vector
}
\description{
DIY your own morandi palette
}
\examples{
x <- c(3, 6, 9, 18, 56)
my_diy_colors <- morandi_diy(my_colors = x, type = c("discrete"))
}
